requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
Django==1.10.2
Fabric3==1.14.post1
idna==2.7
invoke==1.1.1
paramiko==2.4.1
pep8==1.7.0
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
selenium==3.14.0
six==1.11.0
urllib3==1.23
